requirements.txt
Django==1.11.29
djangorestframework==3.14.0
